MultiTemp. classes, imports and tests
