+++ src/smos/smos_l4/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""
 from pkg_resources import get_distribution, DistributionNotFound
 
 try:
@@ -10,5 +11,7 @@ except DistributionNotFound:
 finally:
     del get_distribution, DistributionNotFound
 
-from smos.smos_l4.interface_l4 import SMOSL4Ds, SMOSL4Img
+"""
+
+from smos.smos_l4.interface_l4 import SMOS_L4_Ds, SMOS_L4_Img
 from smos.interface import SMOSTs
