+++ src/smos/smos_l4/interface_l4.py
@@ -22,22 +22,15 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import warnings
 import numpy as np
-import os
-from pygeobase.io_base import ImageBase, MultiTemporalImageBase
-from pygeobase.object_base import Image
-from datetime import timedelta, datetime
 from netCDF4 import Dataset, date2num, num2date
 from smos.grid import EASE25CellGrid
-from smos.smos_ic import dist_name as subdist_name
-from smos import dist_name, __version__
-from collections import OrderedDict
+from smos.interface import SMOSImg, SMOSDs
 
 
-class SMOSL4Img(ImageBase):
+class SMOS_L4_Img(SMOSImg):
     """
-    Class for reading one SMOS L4 CATDS - CESBIO RZSM netcdf image file.
+    Child-class for reading one SMOS L4 CATDS - CESBIO RZSM netcdf image file.
 
     Parameters
     ----------
@@ -74,7 +67,7 @@ class SMOSL4Img(ImageBase):
                  grid=EASE25CellGrid(bbox=None), read_flags=np.linspace(0,1,6,endpoint=True),
                  oper=False, float_fillval=np.nan):
 
-        super(SMOSL4Img, self).__init__(filename, mode=mode)
+        super().__init__(filename, mode=mode)
 
         if parameters is None:
             parameters = []
@@ -95,36 +88,6 @@ class SMOSL4Img(ImageBase):
 
         self.float_fillval = float_fillval
 
-    def get_global_attrs(self, exclude=('history', 'NCO', 'netcdf_version_id', 'contact',
-                                        'institution', 'creation_time')):
-        return {k: v for k, v in self.glob_attrs.items() if k not in exclude}
-
-    def _read_empty(self):
-        """
-        Create an empty image for filling missing dates, this is necessary
-        for reshuffling as img2ts cannot handle missing days.
-
-        Returns
-        -------
-        empty_img : dict
-            Empty arrays of image size for each object parameter
-        """
-        self.image_missing = True
-
-        return_img = {}
-        return_metadata = {}
-
-        try:
-            rows, cols = self.grid.subset_shape
-        except AttributeError:
-            rows, cols = self.grid.shape
-
-        for param in self.parameters:
-            data = np.full((rows, cols), np.nan)
-            return_img[param] = data.flatten()
-            return_metadata[param] = {'image_missing': 1}
-
-        return return_img, return_metadata
 
     def _read_img(self) -> (dict, dict):
         # Read a netcdf image and metadata
@@ -199,139 +162,8 @@ class SMOSL4Img(ImageBase):
 
         return param_img, param_meta
 
-    def read(self, timestamp):
-        """
-        Read a single SMOS image, if it exists, otherwise fill an empty image
 
-        Parameters
-        --------
-        timestamp : datetime
-            Time stamp for the image to read.
-        """
-
-        if timestamp is None:
-            raise ValueError("No time stamp passed")
-
-        try:
-            return_img, return_metadata = self._read_img()
-        except IOError:
-            warnings.warn('Error loading image for {}, '
-                          'generating empty image instead'.format(timestamp.date()))
-            return_img, return_metadata = self._read_empty()
-
-        if self.flatten:
-            self.img = Image(self.grid.activearrlon, self.grid.activearrlat,
-                             return_img, return_metadata, timestamp)
-
-        else:
-            try:
-                shape = self.grid.subset_shape
-            except AttributeError:
-                shape = self.grid.shape
-
-            rows, cols = shape
-            for key in return_img:
-                return_img[key] = np.flipud(return_img[key].reshape(rows, cols))
-
-            self.img = Image(self.grid.activearrlon.reshape(rows, cols),
-                             np.flipud(self.grid.activearrlat.reshape(rows, cols)),
-                             return_img,
-                             return_metadata,
-                             timestamp)
-        return self.img
-
-    def write(self, image, **kwargs):
-        """
-        Write the image to a separate output path. E.g. after reading only
-        a subset of the parameters, or when reading a spatial subset (with a
-        subgrid). If there is already a file, the new image is appended along
-        the time dimension.
-
-        Parameters
-        ----------
-        image : str
-            Path to netcdf file to create.
-        kwargs
-            Additional kwargs are given to netcdf4.Dataset
-        """
-
-        if self.img is None:
-            raise IOError("No data found for current image, load data first")
-
-        if self.img.timestamp is None:
-            raise IOError("No time stamp found for current image.")
-
-        lons = np.unique(self.img.lon.flatten())
-        lats = np.flipud(np.unique(self.img.lat.flatten()))
-
-        mode = 'w' if not os.path.isfile(image) else 'a'
-        ds = Dataset(image, mode=mode, **kwargs)
-
-        ds.set_auto_scale(True)
-        ds.set_auto_mask(True)
-
-        units = 'Days since 2000-01-01 00:00:00'
-
-        if mode == 'w':
-            ds.createDimension('timestamp', None)  # stack dim
-            ds.createDimension('lat', len(lats))
-            ds.createDimension('lon', len(lons))
-
-            # this is not the obs time, but an image time stamp
-            ds.createVariable('timestamp', datatype=np.double, dimensions=('timestamp',),
-                              zlib=True, chunksizes=None)
-            ds.createVariable('lat', datatype='float64', dimensions=('lat',), zlib=True)
-            ds.createVariable('lon', datatype='float64', dimensions=('lon',), zlib=True)
-
-            ds.variables['timestamp'].setncatts({'long_name': 'timestamp',
-                                                'units': units})
-            ds.variables['lat'].setncatts({'long_name': 'latitude', 'units': 'Degrees_North',
-                                           'valid_range': (-90, 90)})
-            ds.variables['lon'].setncatts({'long_name': 'longitude', 'units': 'Degrees_East',
-                                           'valid_range': (-180, 180)})
-
-            ds.variables['lon'][:] = lons
-            ds.variables['lat'][:] = lats
-            ds.variables['timestamp'][:] = np.array([])
-
-            this_global_attrs = \
-                OrderedDict([('subset_img_creation_time', str(datetime.now())),
-                             ('subset_img_bbox_corners_latlon', str(self.grid.bbox)),
-                             ('subset_software', f"{dist_name} | {subdist_name} | {__version__}")])
-            glob_attrs = self.glob_attrs
-            for k in ['ease_global', 'history', 'creation_time', 'NCO']:
-                try:
-                    glob_attrs.pop(k)
-                except KeyError:
-                    continue
-            glob_attrs.update(this_global_attrs)
-            ds.setncatts(glob_attrs)
-
-        idx = ds.variables['timestamp'].shape[0]
-        ds.variables['timestamp'][idx] = date2num(self.img.timestamp, units=units)
-
-        for var, vardata in self.img.data.items():
-
-            if var not in ds.variables.keys():
-                ds.createVariable(var, vardata.dtype, dimensions=('timestamp', 'lat', 'lon'),
-                                  zlib=True, complevel=6)
-                ds.variables[var].setncatts(self.img.metadata[var])
-
-            ds.variables[var][-1] = vardata
-
-        ds.close()
-
-    def read_masked_data(self, **kwargs):
-        raise NotImplementedError
-
-    def flush(self):
-        pass
-
-    def close(self):
-        pass
-
-
-class SMOSL4Ds(MultiTemporalImageBase):
+class SMOS_L4_Ds(SMOSDs):
     """
     Class for reading SMOS L4 CATDS - CESBIO RZSM images in nc format. Images are orgnaised in subdirs
     for each year.
@@ -371,102 +203,15 @@ class SMOSL4Ds(MultiTemporalImageBase):
                  grid=EASE25CellGrid(bbox=None), filename_templ=None,
                  read_flags=np.linspace(0,1,6,endpoint=True), oper=False, float_fillval=np.nan):
 
-        ioclass_kws = {'parameters': parameters,
-                       'flatten': flatten,
-                       'grid': grid,
-                       'read_flags': read_flags,
-                       'oper': oper,
-                       'float_fillval': float_fillval}
-
-        sub_path = ['%Y']
-
         if filename_templ is None:
             filename_templ = self.default_fname_template
 
-        super(SMOSL4Ds, self).__init__(data_path, ioclass=SMOSL4Img,
-                                     fname_templ=filename_templ,
-                                     datetime_format="%Y%m%d",
-                                     subpath_templ=sub_path,
-                                     exact_templ=False,
-                                     ioclass_kws=ioclass_kws)
-
-    def _assemble_img(self, timestamp, mask=False, **kwargs):
-        img = None
-        try:
-            filepath = self._build_filename(timestamp)
-        except IOError:
-            filepath = None
-
-        if self._open(filepath):
-            kwargs['timestamp'] = timestamp
-            if mask is False:
-                img = self.fid.read(**kwargs)
-            else:
-                img = self.fid.read_masked_data(**kwargs)
-
-        return img
-
-    def read(self, timestamp, **kwargs):
-        return self._assemble_img(timestamp, **kwargs)
-
-    def write_multiple(self, root_path, start_date, end_date, stackfile='stack.nc',
-                       **kwargs):
-        """
-        Create multiple netcdf files or a netcdf stack in the passed directory for
-        a range of time stamps. Note that stacking gets slower when the stack gets larger.
-        Empty images (if no original data can be loaded) are excluded here as
-        well.
-
-        Parameters
-        ----------
-        root : str
-            Directory where the files / the stack are/is stored
-        start_date : datetime
-            Start date of images to write down
-        end_date
-            Last date of images to write down
-        stackfile : str, optional (default: 'stack.nc')
-            Name of the stack file to create in root_path. If no name is passed
-            we create single images instead of a stack with the same name as
-            the original images (faster).
-        kwargs:
-            kwargs that are passed to the image reading function
-        """
-        timestamps = self.tstamps_for_daterange(start_date, end_date)
-        for t in timestamps:
-            self.read(t, **kwargs)
-            if self.fid.image_missing:
-                continue
-            if stackfile is None:
-                subdir = os.path.join(root_path, str(t.year))
-                if not os.path.exists(subdir): os.makedirs(subdir)
-                filepath = os.path.join(subdir, os.path.basename(self.fid.filename))
-            else:
-                filepath = os.path.join(root_path, stackfile)
-            print(f"{'Write' if not stackfile else 'Stack'} image for {str(t)}...")
-            self.fid.write(filepath)
-
-    def tstamps_for_daterange(self, start_date, end_date):
-        """
-        return timestamps for daterange,
-        Parameters
-        ----------
-        start_date: datetime.datetime
-            start of date range
-        end_date: datetime.datetime
-            end of date range
-        Returns
-        -------
-        timestamps : list
-            list of datetime objects of each available image between
-            start_date and end_date
-        """
-        img_offsets = np.array([timedelta(hours=0)])
-
-        timestamps = []
-        diff = end_date - start_date
-        for i in range(diff.days + 1):
-            daily_dates = start_date + timedelta(days=i) + img_offsets
-            timestamps.extend(daily_dates.tolist())
-
-        return timestamps
+        super().__init__(data_path, ioclass=SMOS_L4_Img,
+                         parameters=parameters,
+                         flatten=flatten,
+                         grid=grid,
+                         filename_templ=filename_templ,
+                         read_flags=read_flags,
+                         float_fillval=float_fillval,
+                         additional_kws={'oper': oper}
+                         )
